HHH-13725 - Implement ToOne Associations support
