--- hibernate-core/src/main/java/org/hibernate/sql/results/internal/domain/entity/DelayedEntityAssembler.java
@@ -1,13 +1,0 @@
-/*
- * Hibernate, Relational Persistence for Idiomatic Java
- *
- * License: GNU Lesser General Public License (LGPL), version 2.1 or later
- * See the lgpl.txt file in the root directory or http://www.gnu.org/licenses/lgpl-2.1.html
- */
-package org.hibernate.sql.results.internal.domain.entity;
-
-/**
- * @author Andrea Boriero
- */
-public class DelayedEntityAssembler  {
-}
